refactor(folder): moved map component to the the features directory
