+++ src/features/map/components/ZoomControls.tsx
